Add auto redirect to SSO to new login page
